feat: save the genesis hash in home_dir/.nearup/genesis_md5sum and handle records file (#229)
currently nearup computes an md5 hash of the genesis file every time it starts up,
and compares that to the published hash. This leads to a problem where it will unnecessarily
delete the data dir after the first time we start, since neard init rewrites the genesis file
with slight changes: https://github.com/near/nearup/issues/228

So instead of computing the hash every time, we will record the published hash in
{home_dir}/.nearup/genesis_md5sum when initializing the home directory, and we'll read
from that to see if the published version has changed. This also lets us handle the case
where the published genesis is using a separate records file, which is something that may
happen in the near future because it would speed up start up times significantly

At the same time, we'll remove the download_config() function and just use the --download-config
argument to neard init